ConfigureGitalyHooksBin for gitaly-hooks binary
Function ConfigureGitalyHooksBin creates a gitaly-hooks binary
for use in the tests. It does almost the same what ConfigureGitalyHooksBinary
but it is safe to use concurrently from the test. It will
create a binary once in the shared test directory and each
subsequent call to it will copy an existing binary to the proper
location without unneeded compilation. This allows to configure
tests in more granular way. Current approach with configuring per
package is error prone as you may have incorrect test setup because
it is done for the whole package and not the single test.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699