replicator: Do not bump repository generation for PackRefs
Similar in spirit to fc93593cb (praefect: Stop creating replication jobs
on cleanup, 2021-02-03), we're still bumping the repository generation
when creating PackRefs jobs. But given that these only optimize the
repository without changing the user-visible state of the repository,
bumping the repository generation is only going to needlessly cause us
to regard secondaries as out-of-date.

Fix the issue by introducing a separate "PackRefs" change type which
doesn't cause us to bump the generation.