Removal of the config.Config from objectpool package
As we are ready to get rid of the global config.Config
variable this change does exact it for the files in the
objectpool package. It also includes some small refactorings
like adding a setup function, renames and marking functions
as test-helpers.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699