Shortcuts for a quick setup
With addition of the testcfg package the usage pattern of it
pretty strait forward: in a lot of places we create just a
minimum required setup and use it for the tests, or it may
also include a repository creation. That is why two new
functions: Build and BuildWithRepo were added to be used
in those popular cases. It reduces amount of code we need
to support and unifies usage of the testcfg.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699