Makefile: Use per-tool version files
Right now, we determine whether tools need a rebuild based on whether
the Makefile has changed or not. This is kind of a sledge hammer: it
works in most cases, but is overly general. Also, it fails in case the
user specifies a version e.g. via `config.mak` or via the command line.

Refactor this to use per-tool version files and drop the
`Makefile.sha256` hack.