operations: Fix deletion of branches with `refs/heads` prefix
When deleting branches, we currently still have a Ruby bug-for-bug
compatibility layer in place which strips the `refs/heads` prefix of a
branch name in case the user provided it. So if the user intends to
delete `refs/heads/foobar`, we'd try to delete the reference
`refs/heads/foobar` instead of `refs/heads/refs/heads/foobar`.

Now that the Ruby code doesn't exist anymore, let's also drop the
compatibility layer and use sane behaviour here: if the user tells us to
delete a branch `refs/heads/foobar`, we delete a branch
`refs/heads/foobar`.