Makefie: Simplify computation of tool mod directories
In order to find the per-tool go.mod directories, we take the basename
of the tool via `$(notdir $@)`. This is kind of roundabout: we're using
a pattern which matches against the tool's basename already, so we can
instead just match against the stem of the rule.

Simplify the code by using `$*` instead of `$(notdir $@)`.