catfile: Rename `returnWhenDone` to clarify its meaning
The function `returnWhenDone()` reads as if it's only going to return
when the call is done. While that property actually holds, it is not
what the function's intent is. Instead, it returns the catfile process
back to the cache when it's done.

Rename the function to `returnToCacheWhenDone()` to make its intent
clearer.