Makefile: Fix benchmarking repository's URL
The benchmarking repository is supposed to be a not-so-small repository
where it's possible to perform reasonable measurements without
operations concluding too quickly. By accident, it is currently pointing
to the `gitlab-org/gitlab-git.git` repository though, which is a
comparatively small repository. The intent was to have it point to the
GitLab repository though, which is `gitlab-org/gitlab.git`.

Fix the URL to point to the correct repository.