fix flaky TestPerRepositoryElector_Retry
This commit fixes TestPerRepositoryElector_Retry which is currently
flaky. The flaking occurs as the the test returns from the Run loop
by canceling the context. The context is checked in the same select
statement where the retryWait is waited for. In the test, the retry
wait is set to a very low value, so it triggers immediately. The
select statement then receives on both the ctx.Done and time.After
channels and randomly picks one, causing the test to flake due to
calling performFailovers too many times.

This commit fixes the problem by returning an error directly from the
errorHandler set up in the test, which returns directly instead of
hitting the select statement.