hook: Convert pack-objects hook to use `WithGlobalOption()`
The pack-objects hook is using a set of global options. Unfortunately,
it isn't possible to convert this usecase to use non-global arguments
because of the use of the `--shallow-file` flag, which is only
understood at a global level. This commit thus converts the hook to use
`WithGlobalOption()` instead of the `GlobalOption`s parameter.