housekeeping: Refactor tests to use "real" setups
The housekeeping tests currently use a test setup which is quite
detached from reality by only creating empty temporary directories which
serve as git repositories. We're about to change the interface of
`housekeeping.Perform()` to accept a `localrepo.Repo`, and in this
context it doesn't fly anymore to use empty directories.

Improve the test setup by using setups which more closely match an
actual setup by using `testcfg.BuildWithRepo()` to construct the
repository for us.

We're forced to drop one of the tests though which depends on external
setup by the developer in order to create a root-owned repository. This
test is always skipped by default anyway, and I heavily doubt that
anybody ever does the setup procedure to verify it's still working. So
we probably do not loose any real test coverage by dropping the
testcase.