locator: Return proper error code if repo paths do not exist
When retrieving the path of a repository via `GetPath()`, we call
stat(3P) on the calculated path in order to determine whether the
repository exists or not. In case the stat(3P) call errors, we always
return an internal code. This isn't ideal though: while the error may be
because of whatever reason, the most likely reason is that the target
path simply doesn't exist. Returning an internal code is not ideal in
that case.

Fix the issue by returning a `NotFound` code in case the error returned
by stat(3P) is ENOENT.