repository: Perform housekeeping tasks in `OptimizeRepository()`
The `OptimizeRepository` task is all about optimizing repositories, but
in fact it has since grown some functionality which makes sense to
execute via our general housekeeping mechanisms. This is most
importantly the ability to clean up stale configuration of a repo, but
also to clean up empty reference directories.

Furthermore, it makes sense for `OptimizeRepository()` to get hold of
some of the cleanups which our housekeeping package does too. Most
importantly, this includes removal of broken references and stale
lockfiles, both of which are known to break various operations on
repositories. It is thus sensible to start executing these as part of
our regular `OptimizeRepository()` schedule such that they eventually
get repaired.

So let's start calling into the housekeeping procedures when optimizing
the repository. In addition to making available some housekeeping tasks,
this is also a preparatory step to move some functionality from
`OptimizeRepository()` to `housekeeping.Perform()`.