blob: Enable use of bitmap indices when searching LFS pointers
In theory, bitmap indices can help to speed up graph walks by filtering
out irrelevant objects without having to do the actual graph walk. One
such place where they may help is when searching LFS pointers: instead
of identifying all relevant blobs by doing the walk itself, we can
already take a look at the bitmap index to determine all reachable
blobs.

We do not yet use these bitmap indices when doing the graph walk as it
needs to be explicitly enabled for git-rev-list(1). We didn't do so
traditionally because we were relying on the `--object-names` flag as a
heuristic to filter out non-blobs. This heuristic has proven to be not
helpful though and was thus removed in d8a1d23fb (blob: Drop rogue
optimization when reading LFS pointers, 2021-03-02), unblocking our way
to now use bitmap indices.

Enable the use of bitmap indices. Local performance benchmarks are kind
of inconclusive. Executing git-rev-list(1) in the gitlab-org/gitlab
repository directly shows that it should provide a benefit:

    $ time git rev-list --objects --filter=blob:limit=200 --no-object-names --count --all
    2235919

    real	0m13.416s
    user	0m12.015s
    sys	0m1.400s

    $ time git rev-list --objects --filter=blob:limit=200 --no-object-names --count --all --use-bitmap-index
    2235919

    real	0m6.778s
    user	0m5.721s
    sys	0m1.057s

So we're roughly twice as fast. But extending our benchmarks doesn't
show any noticable speedup. Because of this, use of the bitmap index is
hidden behind a new feature flag for now to mark this rather as an
experiment.