repository: Extend race window between ref creation and optimization
When calling `OptimizeRepository()`, one of the things we do is to prune
the repository's reference hierarchy of empty directories. We do so
without any grace period, and as such it is trivial to trigger a race
between a concurrent git process which is just about to create a new
reference and a cleanup which immediately removes the still-empty
directory.

Introduce a grace period so that we only empty directories which are
older than a day. While this doesn't fix the race, it makes it much more
unlikely by giving git processes ample of time to populate their newly
created ref directories.