repository: Replace custom config code with `localrepo.Config`
In order to cleanup leftover configuration of repositories, we've
implemented some custom logic to scan the repository's configuration for
matching keys and remove them. We've since grown a proper interface to
do exactly this via the `localrepo.Config()` interface, so let's convert
the code to use that instead in order to reduce code duplication.