Naming: use "eviction" instead of "expiry"
This updates streamcache comments and (private) struct field names to
use the terms "eviction" and "maximum age" rather than "expiry". In
caching, "expiry" is usually used in context where values may become
stale. In the case of streamcache, we are dealing with immutable
values that cannot become stale. What we are dealing with there is
"eviction".

While we are here, we also update the "status" gauge of the
PackObjectsHook cache to display the maximum age of the cache.