Removal of the config.Config from blob package
As we are ready to get rid of the global config.Config
variable this change does exact it for the files in the
blob package. It also includes some small refactorings
like adding a setup function, renames and marking functions
as test-helpers.

The tests that are dependent on the gitaly-ruby process
now aggregated into a single test that starts gitaly-ruby
process before run with the default configuration. Global
rubyServer variable removed because gitaly-ruby service is
passed as a parameter into the dependent tests together with
the base configuration.

The testhelper.DefaultLocator() was removed because it has
no usages anymore.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699