Removal of the config.Config from commit package
As we are ready to get rid of the global config.Config
variable this change does exact it for the files in the
commit package. It also includes some small refactorings
like adding a setup functions, renames and marking functions
as test-helpers.
The change includes addition of the InitRepoWithWorktreeAtStorage.
It creates an empty repository with a worktree at the
provided storage.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699