repository: Test `force` parameter of `FetchRemote()`
The `FetchRemote()` RPC has a `force` parameter: if it's not set and the
refmap has unforced refspecs, then those unforced refspecs should fail
the fetch and not update the refs. This code is currently untested,
which allowed for a bug when default refspecs are used.

Improve test coverage by adding a bunch of tests which exercise the
`force` parameter and its interplay with forced refspecs.