implement Push on localrepo
This commit implements Push method on localrepo for pushing references
from localrepo to a remote. Push requires refspecs to be specified
always to avoid triggering the default push behavior when refspecs are
not specified which depends on configuration. Push currently force pushes
always as the only call site of the function needs force pushing.