Makefile: Upgrade default git version to v2.31.0
With git v2.31.0 having been released on March 15th, this commit changes
the default version of git to v2.31.0. This will impact both developers
and source installations if they use the "git" target to install git.
Given that downstream distributors do not yet use that target, they're
not impacted, and neither are CI builds as they explicitly override the
git version.