gitaly-git2go: Don't rebase when already at tip
When there are no new commits in the upstream branch since the branch
was created, so the branch you're rebasing is already at the tip of the
upstream branch, do not rebase. There is in fact no need to replay the
commits, and this behavior is also parallel with how git-rebase(1)
handles this case.