Enable gprc-go debug log messages with GRPC_GO_LOG_SEVERITY_LEVEL
https://github.com/grpc/grpc-go#how-to-turn-on-logging mentions that the
`GRPC_GO_LOG_SEVERITY_LEVEL` environment variable can be set to enable
gRPC debugging, but it was not possible to enable this previously. The
log level was always overriden because it was too chatty.

To debug https://gitlab.com/gitlab-org/gitlab/-/issues/297359, we'll
need to be able to see the gprc-go logs for both Gitaly and Praefect.