go.mod: Upgrade to tableflip v1.2.2
Recently, we've had increasing reports about Gitaly not correctly
coming up, where it never starts to accept incoming RPCs. It's unclear
yet what the root cause is, and no reliable reproudcer was found for
this issue yet. But it's clear that something's going on there.

One of the Goroutine dumps provided in this context shows that tableflip
was still running after dozens of minutes, which seems to indicate that
an upgrade went wrong somehow. Digging through changes made in the
tableflip library made me notice commit cae714b (ensure that inherited
files are nonblocking, 2021-01-22), which talks about indefinite hangs
caused by Go clearing `O_NONBLOCK` on exec. While this really is a
shot in the dark and quite unlikely to be the root cause, upgrading our
version of tableflip to get hold of that commit is easy enough to do.

Upgrade the tableflip module to v1.2.2. Most important changes between
the 4baec98 (Add ListenConfig as Option to Upgrader, 2020-05-07) and
v1.2.2 are a bump of the minimum required Go version to v1.14 and the
`O_NONBLOCK` change.