repository: Introduce DefaultRef variable
Gitaly has always assumed the `master` is the default reference for
repositories. This value is also used as fallback. With this change a
future change from `master` to `main` would be easier to achieve.

Further it would allow us to introduce a feature flag or config value
with ease if required.