operations: Add setupOperationClient test helper
In some of the `TestServer_UserCherryPick` cases there was a new context
created in the test. This caused the feature flag not being propagated
to the server, so the test always used the default value for the feature
flag.

This issue was revealed when I flipped the default value of the
`GoUserCherryPick` flag, see [1].

This change introduces a test helper function `setupOperationClient`.
This function takes care of setting up the server, creating a client,
and it will add the metadata to the context.

Using this helper will reduce chances of making the same mistake again.

1. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/3262