git2go: Rename CherryPickCmd author to committer
In the request to do a `CherryPick` with `gitaly-git2go` the fields for
the identity of the committer started with `Author`. This is not
correct, when a commit is cherry-picked the original author is kept.
Only the committer identity and date are overwritten with who/when the
commit is picked.