gitaly-git2go: Preserve cherry-pick author
Prior to this change, gitaly-git2go used the committer as the author and
as the committer. git-cherry-pick(1) on the other hand preserves the
author and only overwrites the committer.

This change makes gitaly-git2go to behave the same and retains the
original author but sets the committer as specified in the command
request.

Tests were added check these commit properties, but because git(1)
stores dates as Unix time + offset as helper function is added to
compare two `git2go.Signature`s.