Add ConfigureGitalyGit2GoBin
This change adds ConfigureGitalyGit2GoBin that
compiles or copies gitaly-git2go executable into
proper place according to the passed in cfg.
The buildCommand is extended with a new parameter
testing.TB to pass it into the MustRunCommand so
it properly reports in case of a failure.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699