Remove config.Config from conflicts package
As we are ready to get rid of the global config.Config
variable this change does exact it for the files in the
conflicts package. It also includes some small refactoring
like adding setup functions, renames and marking functions
as test-helpers.
All ruby sidecar dependent tests are assembled into a single one
that uses the local instance of the ruby service and
the global RubyServer variable is removed.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699