Removal of config.Config from diff package
As we are ready to get rid of the global config.Config
variable this change does exact it for the files in the
diff package. It also includes some small refactoring
like adding a setup functions, renames and marking functions
as test-helpers.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699