git: Document downstream users which need update on version changes
As Gitaly is starting to become the sole distributor of the git version,
there will be less sites that need updating of the git version when the
minimum required version changes. No matter what, some will still need
updates though, while others go away.

Next up is gitlab-org/omnibus-gitlab, which is about to be converted to
use Gitaly's provided git version. It always used to have a very helpful
list of places which need updating when bumping the git version though,
which is also going away now. So let's adopt it into Gitaly such that it
does not get lost as a checklist to see whether any location was missed.