Removal of config.Config from operations package
As we are ready to get rid of the global config.Config
variable this change does exact it for the files in the
operations package. It also includes some small refactoring
like adding a setup functions, renames, marking functions
as test-helpers. The tests dependent on usage of the ruby-sidecar
assembled into a single TestWithRubySidecar test and runs
as separate sub-tests re-using same ruby-sidecar instance.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699