WIP Feature flag rollout docs: correct OnByDefault and removal examples
The change I made in 938a75444 (Feature flag rollout docs: add
examples of Go/Ruby code removal, 2021-02-25) didn't make much
sense. The link to the "on by default" didn't actually link to such a
MR.

This would be ideal, but in !3033 I made the stupid mistake of mixing
up "delete" and "create" in the subject.

So let's have this wait, will rewrite this once I have
UserUpdateBranch landed with something more sensible I can link to as
an example.