Makefile: Switch to git-checkout(1) to allow oldish host git
We're using git-switch(1) to switch between different revisions when
building either the git or libgit2 targets. git-switch(1) may not be
available on some systems given that it's only been introduced with git
v2.23.0, so building those targets on such systems is broken.

Fix the issue by using the venerable git-checkout(1) command instead.
For our usecase, they're doing the same anyway.