implement multiplexed connections from Praefect to Gitaly
This commit integrates connection multiplexing in Gitaly and Praefect.
Praefect dials both regular and multiplexed connections to Gitaly.
The `gitaly_connection_multiplexing` feature flag is used to toggle
which connection Praefect uses to send the requests to the Gitaly.
The multiplexed connections are not health checked separately as they
are not tied to the leader election logic. Connection multiplexing is
only used if `sql` election strategy is configured, as the implementation
is only done there. Replicaton queue operations are do not go overm multiplexed
connections as it is not possible to feature flag them. This should be
sufficient for performance testing, proper integration follows with the voting
logic changes if the performance is acceptable.