maintenance: Introduce new random filesystem walker
In order to get a random distribution of optimized repositories, the
optimization maintenance job does a random walk through all storage
repositories. This random walk is implemented ad-hoc via a recursive
function which is deeply married with the optimization calls. This makes
it hard to extend, reason about and test.

To improve the situation, this commit thus introduces a generic random
filesystem walker implementation which has the same semantics as the
current implementation: given a directory hierarchy, the walker will go
down this hierarchy depth-first and, for each directory it encounters,
process its entries in a randomized order.

The new walker is not yet used.