gitaly-git2go: Ignore nsecs and timezone in tests
git(1) stores commits with a one second precision, and it stores the
timezone offset, but not it's name.

This change sets the nanosecond offset for `DefaultAuthor` to zero, and
the timezone name to an empty string. This makes it possible to use
`require.Equal` to compare a given `git2go.Signature` with a signature
deserialized from an actual commit.

Fixes https://gitlab.com/gitlab-org/gitaly/-/issues/3534.