Refactoring of the ListenGitalySSHCalls
The ListenGitalySSHCalls function used only in one place
so we move it closer to where it is used.
The function also changed to omit modification of the
passed in config.Cfg parameter, so we don't need to override
it after setup is done.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699