Do not return error from streamcache.New
The way we wire up gRPC services, it is awkward if a gRPC service
server constructor can return an error. This commit refactors
streamcache.New and streamcache.newFilestore to never return an error.
The one thing that could return an error now happens lazily.

Because there is now more synchronization in filestore, this change
includes a refactor where we change to a single mutex for the whole
struct.