Support variable replication factor in RepositoryReplicas
This commit changes RepositoryReplicas to support variable replication
factor and repository specific primaries.

RepositoryReplicas is an RPC that returns the checksums of the
repository's replicas and which one is the primary. Currently the RPC
only supports virtual storage scoped primaries due to use of GetShard.
This commit plugs in PrimaryGetter so the RPC can use repository specific
primary if they are enabled.

Note that the RPC fails if one of the secondaries does not actually
contain the repository. Due to this behavior, the assignments store is
also injected to only get the checksum from the assigned host storages of
the repository. This is done to avoid failures as the repository is not
necessarily present on every storage with variable replication factor. With
that in mind, some unassigned nodes may actually contain the repository but
we do not return their checksums.