Remove NodeManager dependency from DiskStats
Praefect's DiskStats gathers disk statistics from all of the storages
configured in Praefect. It currently uses NodeManager's GetShard to
get the storage names and connections. This commit removes the dependency
so the function works the same with virtual storage scoped primaries and
with repository specific primaries configured. It now uses an injected
set of connections that connections for each storage configured.