Remove NodeManager dependency from ServerInfo
This commit removes the NodeManager dependecny from ServerInfo so
it works equally with virtual storage scoped primaries or repository
specific primaries configured.

The RPC is fairly eccentric. It tries to give an illusion of a single
storage responding, which does not really suit well with the multiple
storages sitting behind a Praefect. It previously relayed the information
from the virtual storage's primary. With repository specific primaries,
none of the storages have a special status. As such, we'll just return the
response from a random storage from now on. Given the primary might change
due to a failover, the client's can't rely on the returned response anyway.