git: Split command options into separate file
Over the years, we have grown quite a lot of options which can be passed
to git commands. This has made the actual implementation of git commands
quite hard to read given that the code file where it's hosted does so
many different things now.

Split out command options into a separate file to fix this and help
readability.