git: Rename `gitCommand` to `commandDescription`
The `gitCommands` array hosts descriptions about all the git commands
we're running in Gitaly's codebase. Its naming is somewhat awkward
though because it clashes with the `Cmd` interface and its `SubCmd`
respectively `SubSubCmd` implementations. By only taking a look at their
names, it's not obvious what the scope of each of these is and how they
interact with each other.

Rename the `gitCommands` array and its `gitCommand` structure to
`commandDescriptions` respectively `commandDescription`. Because this is
what it is: it describes git commands, how we need to invoke them and
how they behave.