git: Move argument assembly into command description
Assembly of command args is currently hosted next to the `Cmd` interface
and its implementations. While this isn't wrong per se, it also relies
on the command descriptions such that it knows how options can be
passed.

This commit thus moves over assembly of command arguments to be a
receiver function for the command descriptions. This allows us to easily
extend command descriptions to implement custom per-command validations
of arguments in a subsequent commit.