git: Allow per-command validation of positional args
Currently, all git commands use the same logic to validate global,
positional and post-separator arguments. This is inflexible, because
different commands use different syntax and also have different
behaviour. Until now, this hasn't bitten us yet and has served its
purpose of disallowing injection of command line options. But for some
commands, the existing sets of rules bar us from using some features.

Introduce a new optional function for command descriptions which allow
us to override validation of positional arguments per command. This can
be used in arbitrary ways: either to strengthen validation and assert
that certain kinds of arguments cannot ever be passed, or to weaken
validation in case we know that a set of arguments is safe to be
accepted as positional argument.

A user of this is going to be introduced in the following commit.