git: Allow git-rev-list(1) to mix revs and pseudo-revs
The git-rev-list(1) command is used to compute a list of objects based
on a set of revisions passed in by the user. Next to passing normal
revisions, it also allows pseudo-revisions like "--not" or "--all" which
have some special logic and may influence normal revisions which follow
after the given pseudo-revision. But because they start with a leading
dash and we refuse positional arguments with dashes, we haven't been
able to use their full potential yet.

With the new per-command validation of positional arguments, this
shortcoming can now easily be fixed: while we keep normal validation of
positional arguments in place, we do override it for "--all" and "--not"
such that they're allowed in positional arguments when git-rev-list(1)
executes. They're harmless, and being able to mix them with normal
revisions will allow better interfaces for us in the future.